+++ Sora/PackageInfo.swift
@@ -1,7 +1,7 @@
 /// :nodoc:
 public enum SDKInfo {
     // Sora iOS SDK のバージョンを定義する
-    public static let version = "2022.6.0"
+    public static let version = "2023.1.0"
 }
 
 /**
@@ -9,16 +9,16 @@ public enum SDKInfo {
  */
 public enum WebRTCInfo {
     /// WebRTC フレームワークのバージョン
-    public static let version = "M105"
+    public static let version = "M112"
 
     /// WebRTC フレームワークのコミットポジション
-    public static let commitPosition = "0"
+    public static let commitPosition = "1"
 
     /// WebRTC フレームワークのメンテナンスバージョン
     public static let maintenanceVersion = "0"
 
     /// WebRTC フレームワークのソースコードのリビジョン
-    public static let revision = "dc5cf31cad576376abd3aa6306169453cfd85ba5"
+    public static let revision = "18a31880e3037f9938581c1210d2d73b9685f4aa"
 
     /// WebRTC フレームワークのソースコードのリビジョン (短縮版)
     public static var shortRevision: String {